requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.32.0
jinja2>=3.1.4
httpx>=0.28.0
python-dotenv>=1.0.0
python-multipart>=0.0.12
sqlalchemy>=2.0.0
alembic>=1.13.0
